requirements/common.txt
tornado==5.1
python-dotenv==0.9.1
aiomysql==0.0.19
schema==0.6.8
bcrypt==3.1.4
PyJWT==1.6.4